Call Reference

Constructor

ChemicalFormula.Formula — Type

Formula(formula[, composition], charge=0, name=nothing)

Represent a chemical formula with an electrical charge and an optional name. The composition is automatically determined from the specified formula. Compounds can be grouped with parentheses, coordinating molecules can be annotated with a *. Elements are represented by Symbols.

Examples

julia> Formula("H2O", "water")
Formula("H2O", Dict{Symbol, Int32}(:H => 2, :O => 1), 0, "water")
julia> Formula("SO4", -2)
Formula("SO4", Dict{Symbol, Int32}(:S => 1, :O => 4), -2, nothing)
julia> Formula("Fe(CN)6*5H2O")
Formula("Fe(CN)6*5H2O", Dict{Symbol, Int32}(:N => 6, :Fe => 1, :H => 10, :O => 5, :C => 6), 0, nothing)

String formatting

ChemicalFormula.unicode — Function

unicode(formula, includecharge=true, form="formula")

Give a formula written in unicode characters. includecharge determines whether the electronic charge should be given, defaults to true. form has to be one of "formula" (the formula that was given to construct the Formula object), "hill" or "hillformula" for Hill notation or "sum" or "sumformula" for a collapsed sum formula, defaults to "formula".

ChemicalFormula.latex — Function

latex(formula, includecharge=true, form="formula")

Give a formula written in LaTeX, using the mhchem package. includecharge determines whether the electronic charge should be given, defaults to true. form has to be one of "formula" (the formula that was given to construct the Formula object), "hill" or "hillformula" for Hill notation or "sum" or "sumformula" for a collapsed sum formula, defaults to "formula".

ChemicalFormula.sumformula — Function

Give a collapsed sum formula with no special ordering and without charge as unformatted text.

ChemicalFormula.hillformula — Function

Give a sum formula in Hill notation without charge as unformatted text.

ChemicalFormula.textcharge — Function

Return formatted String of the formula charge (like 4+, 3-, +, ...).

Mass and weight related functions

ChemicalFormula.formulamass — Function

Return the mass of one formula unit in unified atomic mass units $u$.

ChemicalFormula.formulaweight — Function

Return the molar mass of the formula in $g mol⁻¹$.

ChemicalFormula.massfractions — Function

Return the mass fraction for each element of the formula in a element => fraction Dict.

Utility functions

ChemicalFormula.ischarged — Function

Return whether the formula carries an electrical charge.

ChemicalFormula.isradioactive — Function

Return whether the formula is radioactive, i.e. contains radioactive elements.